requirements.txt
# Core
torch>=2.2.0
torchvision>=0.17.0
numpy>=1.24.0
pillow>=10.0.0
scikit-learn>=1.3.0
matplotlib>=3.8.0
seaborn>=0.13.0
tqdm>=4.66.0
pandas>=2.1.0
joblib>=1.3.0

# Quantum
pennylane>=0.33.0

# Optional UI
flask>=3.0.0
